Keep DOC, STET, and Passages better separated
