+++ backend/document/entrypoints/stet/routes.py
@@ -5,7 +5,8 @@ from fastapi import APIRouter
 import celery.states
 from celery.result import AsyncResult
 from document.config import settings
-from document.domain import document_generator, resource_lookup
+from document.stet import document_generator
+from document.domain import resource_lookup
 from document.stet import model
 
 from fastapi import HTTPException, status
@@ -76,7 +77,9 @@ async def generate_docx_document(
         )
     except HTTPException as exc:
         raise exc
-    except Exception as exc:  # catch any exceptions we weren't expecting, handlers handle the ones we do expect.
+    except (
+        Exception
+    ) as exc:  # catch any exceptions we weren't expecting, handlers handle the ones we do expect.
         logger.exception(
             "There was an error while attempting to fulfill the document "
             "request. Likely reason is the following exception:"
